+++ spec/rubocop/cop/style/trailing_blank_lines_spec.rb
@@ -13,6 +13,11 @@ describe Rubocop::Cop::Style::TrailingBlankLines, :config do
       expect(cop.offenses).to be_empty
     end
 
+    it 'accepts an empty file' do
+      inspect_source(cop, '')
+      expect(cop.offenses).to be_empty
+    end
+
     it 'registers an offense for multiple trailing blank lines' do
       inspect_source(cop, ['x = 0', '', '', '', ''])
       expect(cop.offenses.size).to eq(1)